Split sticker types
