+++ packages/types/src/discord/gateway.ts
@@ -4,7 +4,7 @@
  *  - https://discord.com/developers/docs/events/gateway-events
  */
 
-import type { DiscordAvatarDecorationData, DiscordRole, DiscordSticker, DiscordUser } from '../discord.js'
+import type { DiscordAvatarDecorationData, DiscordRole, DiscordUser } from '../discord.js'
 import type { GatewayEventNames } from '../shared.js'
 import type { DiscordApplication } from './applications.js'
 import type { AutoModerationTriggerTypes, DiscordAutoModerationAction } from './autoModeration.js'
@@ -14,6 +14,7 @@ import type { DiscordIntegration, DiscordMember, DiscordMemberWithUser, DiscordU
 import type { TargetTypes } from './invites.js'
 import type { DiscordReactionType } from './messages.js'
 import type { DiscordSoundboardSound } from './soundboards.js'
+import type { DiscordSticker } from './stickers.js'
 
 /** https://discord.com/developers/docs/events/gateway#list-of-intents */
 export enum GatewayIntents {
