+++ packages/types/src/index.ts
@@ -17,6 +17,7 @@ export * from './discord/reference.js'
 export * from './discord/skus.js'
 export * from './discord/soundboards.js'
 export * from './discord/stageInstances.js'
+export * from './discord/stickers.js'
 export * from './discord/webhookEvents.js'
 export * from './discordeno.js'
 export * from './shared.js'
